auth-microservice/app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
requests==2.28.2
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.24.1